widget/tests/test_plugin_scroll_invalidation.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for plugin child widgets not being invalidated by scrolling</title>
  <script type="text/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="text/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css"
          href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>
<body onload="initialize()">
<p id="display">
  <iframe id="i" src="plugin_scroll_invalidation.html"
   width="50" height="50" scrolling="no"></iframe>
</p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>

<script class="testbody" type="application/javascript">
SimpleTest.waitForExplicitFinish();

var scrolling;
var scrolling_plugins = [];
var paint_waiter;
var last_paint_count;

function initialize() {
  scrolling = document.getElementById("i").contentWindow;
  scrolling_plugins = scrolling.document.querySelectorAll("embed.scrolling");
  paint_waiter = scrolling.document.getElementById("paint-waiter");

  // Plugins are not instantiated until after reflow
  waitForPlugin(startTest);
}

function startTest() {
  scrolling.scrollTo(50, 45);
  is(paint_waiter.getPaintCount(), 0, "zero-sized plugin not painted");

  waitForPaint(scrollAround);
}

function scrollAround() {
  var paints = getPaintCounts();

  for (var i = 0; i < paints.length; ++i) {
    isnot(paints[i], 0, "embed " + scrolling_plugins[i].id + " is painted"); 
  }

  last_paint_count = sum(paints);

  scrolling.scrollBy(-5, 5);
  scrolling.scrollBy(5, 5);
  scrolling.scrollBy(5, -5);
  scrolling.scrollBy(-5, 5);

  scrolling.scrollTo(45, 45);
  scrolling.scrollBy(10, 0);
  scrolling.scrollBy(0, 10);
  scrolling.scrollBy(-10, 0);
  scrolling.scrollBy(0, -10);

  waitForPaint(done);
}

function done() {
  is(sum(getPaintCounts()), last_paint_count, "no paint on scroll");
  SimpleTest.finish();  
}

function waitForPlugin(func) {
  if (paint_waiter.getPaintCount) {
    func();
    return;
  }
  setTimeout(function() { waitForPlugin(func); }, 0);
}

function waitForPaint(func) {
  paint_waiter.last_paint_count = paint_waiter.getPaintCount();
  paint_waiter.style.left = scrolling.scrollX + "px";
  paint_waiter.style.top = scrolling.scrollY + "px";
  paint_waiter.style.width = "1px";
  paint_waiter.style.height = "1px";
  waitForPaintHelper(func);
}

function waitForPaintHelper(func) {
  if (paint_waiter.getPaintCount() != paint_waiter.last_paint_count) {
    // hide the paint waiter
    paint_waiter.style.width = "0px";
    paint_waiter.style.height = "0px";
    setTimeout(func, 0);
    return;
  }
  setTimeout(function() { waitForPaintHelper(func); }, 100);
}

function getPaintCounts() {
  var result = [];
  for (var i = 0; i < scrolling_plugins.length; ++i) {
    result[i] = scrolling_plugins[i].getPaintCount();
  }
  return result;
}

function sum(array) {
  var result = 0;
  for (var i = 0; i < array.length; ++i) {
    result += array[i];
  }
  return result;
}

</script>
</body>
</html>
